% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get all data from phillymetal.net}
\usage{
get_data(upcoming_shows_only = FALSE)
}
\arguments{
\item{upcoming_shows_only}{\code{boolean} \code{TRUE}: only shows on current date or in the future; \code{FALSE}: all shows that were listed, past, present, or future.}
}
\value{
\code{data.frame} with columns
\itemize{
\item \code{added}: date added to website
\item \code{url}: URL of show
\item \code{show_date} date of show
\item \code{description}: band lineup
\item \code{venue}: venue
\item \code{id}: database key; unique id
\item \code{validated}: for user-submitted data; if TRUE, show is listed. if FALSE, waiting approval or rejected
}
}
\description{
Get all data from phillymetal.net
}
\examples{
all_data <- get_data()

}
